Amend styling of homepage and titles of recipe pages
I felt like the previous design was not an appropriate demonstration of my current CSS studies.
In order to reflect this better I have begun the process of updating the whole website design to
something more modern. I have completed this task for the headings of each page in the website, and
will continue to do this for the contents of each page too.