Update stylings for headings in recipe pages
By applying the recipe-content class to the body directly, I had created disunity
in the website design, as it meant that the default top margin was changed to 0
on the sub-pages but not for the homepage. I created divs in order to counter
this. I also added full stops to titles on recipe pages for the sake of
consistency.